pyproject.toml
[tool.poetry]
name = "ton-connect"
version = "0.1.0"
description = "Python SDK for TON Connect"
authors = ["Yury Sokov <me@yurzs.dev>"]
readme = "README.md"
homepage = "https://github.com/yurzs/ton-connect"
documentation = "https://github.com/yurzs/ton-connect"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11"
aiohttp = "^3.9.5"
pynacl = "^1.5.0"
aiohttp-sse-client = "^0.2.1"
certifi = "^2024.6.2"
pymongo = "^4.10.1"
pydantic = "^2.10.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.7"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = [ "E", "F", "W", "I", "N", "UP", "YTT", "ANN", "S", "B",]
ignore = [ "W191", "E111", "E114", "E117", "ANN101", "ANN102", "ANN002", "ANN003",]
